--- src/dao/filemanagers/carts.json
@@ -1,1 +1,0 @@
-[]
