--- src/dao/filemanagers/products.json
@@ -1,1 +1,0 @@
-[{"title":"Producto 1","description":"Descripción del Producto 1","code":"P001","price":19.99,"status":true,"stock":50,"category":"Electrónica","thumbnails":["imagen1.jpg","imagen2.jpg"],"id":"zARO9"},{"title":"Producto 2","description":"Descripción del Producto 2","code":"P001","price":45,"status":true,"stock":53,"category":"Hogar","thumbnails":["imagen1.jpg","imagen2.jpg"],"id":"HsKCf"},{"title":"Producto 3","description":"Descripción del Producto 3","code":"P003","price":9.99,"status":false,"stock":0,"category":"Hogar","thumbnails":["imagen5.jpg","imagen6.jpg"],"id":"npEe_"},{"title":"Producto 4","description":"Descripción del Producto 4","code":"P004","price":49.99,"status":"Disponible","stock":15,"category":"Electrónica","thumbnails":["imagen7.jpg","imagen8.jpg"],"id":"3m26Z"},{"title":"Producto 5","description":"Descripción del Producto 5","code":"P005","price":14.99,"status":"Disponible","stock":40,"category":"Ropa","thumbnails":["imagen9.jpg","imagen10.jpg"],"id":"Vpaj2"}]
