+++ src/routes/cart.routes.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import Cart from "../DAO/dbManager/carts.js";
+import Cart from "../dao/carts.js";
 
 const cartRouter = Router(); //Crear enrutador
 const cartManager = new Cart();
