+++ src/routes/product.routes.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import Product from "../DAO/dbManager/products.js";
+import Product from "../dao/products.js"
 
 const productRouter = Router();
 const productManager = new Product();
