+++ src/views/chats.handlebars
@@ -16,6 +16,6 @@
     <button type="submit">Ingresar Mensaje</button>
 </form>
     </div>
-    <script src="./public/js/chat.js"></script>
+    <script src="/js/chat.js"></script>
 </body>
 </html>
