+++ src/views/login.handlebars
@@ -10,4 +10,4 @@
     </form>
     <p>¿No estás registrado? <a href="/register">Regístrate aquí</a></p>
 </div>
-<script type="module" src="/login.js"></script>
+<script type="module" src="/js/login.js"></script>
