+++ src/views/register.handlebars
@@ -21,4 +21,4 @@
     </form>
     <p>¿Ya tienes una cuenta? <a href="/login">Ingresa aquí</a></p>
 </div>
-<script type="module" src="/register.js"></script>
+<script type="module" src="/js/register.js"></script>
